[#1110] improvement(coordinator): introduce pluggable remote storage config format (#1329)
Introduce the new `DynamicClientConfigService` to provide the pluggable reader for config file multi format.
And it also is compatible with the original dynamic config format.

This PR introduces the new yaml parser, which is recommended in the future, the storage's hadoop conf could be xml style or k1:v1 . The example as follows.

```yaml
rssClientConf:
  k1: v1
  k2: v2

remoteStorageInfos:
  hdfs://a-ns01: |+
    <configuration>
        <property>
          <name>k1</name>
          <value>v1</value>
        </property>

        <property>
          <name>k2</name>
          <value>v2</value>
        </property>
    </configuration>

  hdfs://x-ns01:
    k1: v1
    k2: v2
```

Fix: #1110

No.

1. UTs